refactor : Participant가 공통 로직을 가지도록 수정
